[Bronze V] Title: 이상한 기호, Time: 124 ms, Memory: 14244 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.java
@@ -7,8 +7,8 @@ public class Main {
     public static void main(String[] args) throws IOException {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
         StringTokenizer st = new StringTokenizer(br.readLine());
-        int A = Integer.parseInt(st.nextToken());
-        int B = Integer.parseInt(st.nextToken());
+        long A = Integer.parseInt(st.nextToken());
+        long B = Integer.parseInt(st.nextToken());
 
         long result = (A + B) * (A - B);
         System.out.println(result);
